+++ source/shaders/GlitchShader.cpp
@@ -15,7 +15,7 @@ void ShaderManager::drawGlitchSkew(const RT& rt, C3D_RenderTarget* dest, float s
     for (int i = 0; i < numStrips; i++) {
         float offsetY = i * stripH;
         // Pseudo-random glitch offset
-        float randomVal = fmod(sinf(time * 12.3f + i * 8.4f) * 43758.5453f, 1.0f);
+        float randomVal = fmodf(sinf(time * 12.3f + i * 8.4f) * 43758.5453f, 1.0f);
         float waveOffset = (randomVal > 0.95f) ? (sinf(time * speed) * strength * randomVal) : 0.0f;
         
         Tex3DS_SubTexture& tempSubtex = tempSubtexs[i % 512];
